+++ src/Rocks.IntegrationTests/AbstractClassGenericMethodTests.cs
@@ -165,7 +165,7 @@ public static class AbstractClassGenericMethodTests
 
 		var mock = expectations.Instance();
 
-		Assert.That(mock.BarReturn<string>, Throws.TypeOf<MockException>());
+		Assert.That(mock.BarReturn<string>, Throws.TypeOf<NoReturnValueException>());
 	}
 
 	[Test]